Refactor drive/mag logic, PWM API, and types
Move magnetometer complementary-filter state into SimpleMelt class and expose tuning params (comp_xy, comp_z, alpha_xy, alpha_z). Simplify arcadeStateUpdate by using local fwd/turn, preserve input signs, and scale outputs with new arcade_power field instead of hardcoded 0.1.

Add ESP Arduino Core v3 compatibility in OneShot125: use ledcAttach/ledcWrite(pin, ...) when core >= 3, and retain ledcSetup/ledcAttachPin/ledcWrite(channel, ...) for older cores. Narrow OneShot125.pin to uint8_t. Remove obsolete keywords.txt and a few unused utility prototypes from SimpleMelt.h.

Reasons: fix heading/magnitude logic, make arcade output tunable, and support newer ESP core PWM API while keeping backwards compatibility.

+++ src/SimpleMeltUtility.h
@@ -27,7 +27,7 @@ class OneShot125 {
         void setReversed(bool isReversed) {reversed = isReversed; };
     private:
         uint8_t channel;
-        uint16_t pin;
+        uint8_t pin;
         bool reversed = false;
         static const int PWM_FREQUENCY = 400; // Hz
         static const int PWM_RESOLUTION = 12; // bits
